(PC-21625) refactor(search): export bodysearch component
